ASMC, flags, and saving flags.
Gave a use to WMDataStruct's node flag bitfield bit 2 to indicate the node has a base available.

Used redundant bits in WMSaveStruct to save aforementioned flags. This'll allow both ExModularSave users and vanilla to (mostly) painlessly add patch. Except those who use a custom WMSaver and/or WMLoader function, but those people probably know how to deal with that.

Added ASMCs for setting, unsetting and inverting availability of base for a given node.